requirements.txt
# ============================================================
# requirements.txt — Python dependencies from Dockerfile
# Base image: nvidia/cuda:12.9.1-cudnn-runtime-ubuntu24.04
# Python: 3.11 (via Miniforge / conda-forge)
# ============================================================

# ----------------------------
# Jupyter & LSP (conda-forge)
# ----------------------------
# jupyterhub==5.2.1
# jupyterlab==4.3.4
# jupyter-lsp
# python-lsp-server

# ----------------------------
# Scientific Stack (conda-forge)
# Install these via conda/mamba for best compatibility:
#   mamba install -c conda-forge numpy pandas scipy scikit-learn ...
# ----------------------------
numpy>=2.1
pandas>=2.2.2
scipy>=1.14
scikit-learn>=1.5
xarray>=2024.6
netcdf4>=1.7
h5netcdf>=1.3
dask>=2024.5
zarr>=2.18
bottleneck>=1.4
cartopy>=0.23
shapely>=2.0
pyproj>=3.6
esmf==8.6.0          # conda-forge only
esmpy==8.6.0         # conda-forge only
xesmf>=0.8.7
matplotlib==3.10.8
seaborn==0.13.2
packaging<=24.2
hydra-core>=1.3.0
omegaconf>=2.3.0
# ----------------------------
# pip-installable packages
# ----------------------------

# Deep Learning — TensorFlow (GPU via CUDA 12.x / cuDNN 9)
tensorflow[and-cuda]==2.18

# Deep Learning — PyTorch + TorchVision (CUDA 12.8 wheel)
# Install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
torch
torchvision